+++ extensions/clickup/src/capture.tsx
@@ -1,17 +1,20 @@
-import { Form, ActionPanel, Action, showToast, Toast, popToRoot } from "@raycast/api";
+import { Form, ActionPanel, Action, showToast, Toast, popToRoot, Icon } from "@raycast/api";
 import { ClickUpClient } from "./utils/clickUpClient";
 import { TaskItem } from "./types/tasks.dt";
 import preferences from "./utils/preferences";
 import { FormValidation, useForm } from "@raycast/utils";
+import { useList } from "./hooks/useList";
 
 interface FormValues {
   name: string;
   description: string;
   dueDate: Date | null;
   priority: string;
+  status: string;
 }
 
 export default function QuickCapture() {
+  const { isLoading, list } = useList(preferences.listId);
   const { itemProps, handleSubmit } = useForm<FormValues>({
     async onSubmit(formValues) {
       const toast = await showToast({
@@ -24,6 +27,7 @@ export default function QuickCapture() {
           ...(formValues?.description && { description: formValues.description }),
           ...(formValues?.dueDate && { due_date: new Date(formValues.dueDate).getTime() }),
           ...(formValues?.priority && { priority: formValues.priority }),
+          ...(formValues?.status && { status: formValues.status }),
         });
 
         // Ensure the user sees root search when they re-open Raycast.
@@ -35,7 +39,7 @@ export default function QuickCapture() {
         toast.title = "Task created successfully";
 
         popToRoot();
-      } catch (error) {
+      } catch {
         toast.title = "Something went wrong";
         toast.message = "Please try again";
         toast.style = Toast.Style.Failure;
@@ -51,6 +55,7 @@ export default function QuickCapture() {
 
   return (
     <Form
+      isLoading={isLoading}
       actions={
         <ActionPanel>
           <Action.SubmitForm title="Submit" onSubmit={handleSubmit} />
@@ -66,6 +71,16 @@ export default function QuickCapture() {
         <Form.Dropdown.Item value="2" title="High" icon="🟠" />
         <Form.Dropdown.Item value="3" title="Normal" icon="🟡" />
         <Form.Dropdown.Item value="4" title="Low" icon="🟢" />
+      </Form.Dropdown>
+      <Form.Dropdown title="Status" storeValue throttle {...itemProps.status}>
+        {list?.statuses.map((status) => (
+          <Form.Dropdown.Item
+            key={status.status}
+            icon={{ source: Icon.CircleFilled, tintColor: status.color }}
+            title={status.status.toUpperCase()}
+            value={status.status}
+          />
+        ))}
       </Form.Dropdown>
     </Form>
   );