+++ extensions/clickup/src/hooks/useClickUp.ts
@@ -11,6 +11,7 @@ export default function useClickUp<T>(endpoint: string, { apiVersion }: { apiVer
     async () => {
       try {
         const response = await ClickUpClient<T>(endpoint, "GET", undefined, undefined, apiVersion);
+        console.log(response);
         return response.data;
       } catch (error) {
         const result = error as AxiosError<ErrorResult>;
@@ -23,7 +24,7 @@ export default function useClickUp<T>(endpoint: string, { apiVersion }: { apiVer
       failureToastOptions: {
         title: "ClickUp Error",
       },
-    }
+    },
   );
   return { isLoading, data };
 }